[Refactoring] Se traslada GetAutoresDeArticulo y GetInformacionArticulo a InformacionArticuloDBHandler
